fix(proxy): make ampcode passthrough fully transparent
The /api/* and /v0/management/* catch-all proxy was rewriting the amp
CLI's own headers (stripping Authorization/X-Api-Key, injecting byokey's
amp token) and buffering the entire upstream response before returning
it. Both broke the amp `read` tool: the rewritten auth caused thread
operations to run as the wrong principal, and buffering prevented the
SSE turn-continuation from streaming back to the client.

Forward the amp CLI's request verbatim (only hop-by-hop + host stripped
per RFC 7230) and stream the response body back. /api/provider/* is
unaffected — byokey still owns AI traffic. The forward_headers
middleware is deleted since no other route needed it.
