+++ crates/proxy/src/middleware/mod.rs
@@ -1,4 +1,3 @@
 //! Axum middleware layers for the proxy.
 
 pub mod dump;
-pub mod forward;
